added status

+++ controllers/feed.js
@@ -1,3 +1,4 @@
+const User = require('../models/user');
 const path = require('path');
 const fs = require('fs');
 const Post = require('../models/post');
@@ -83,21 +84,31 @@ exports.postPosts = (req, res, next) => {
     const imageUrl = req.file.path;
     const title = req.body.title;
     const content = req.body.content;
+    let creator;
     const post = new Post({
         title: title,
         content: content,
         imageUrl: imageUrl,
-        creator: {
-            name: 'Leo'
-        }
+        creator: req.userId
     });
     post
         .save()
         .then(result => {
-            console.log(result);
+            return User.findById(req.userId)
+        })
+        .then(user => {
+            creator = user;
+            user.posts.push(post);
+            return user.save();
+        })
+        .then(() => {
             res.status(201).json({
                 message: 'Post created',
-                post: result
+                post: post,
+                creator: {
+                    _id: creator._id,
+                    name: creator.name
+                }
             });
         })
         .catch(err => {
@@ -158,6 +169,11 @@ exports.updatePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorized');
+                error.statusCode = 403;
+                throw error;
+            }
             if (imageUrl !== post.imageUrl) {
                 clearImage(post.imageUrl);
             }
@@ -192,17 +208,28 @@ exports.deletePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+            if (post.creator.toString() !== req.userId) { //don't forget to refactor later
+                const error = new Error('Not authorized');
+                error.statusCode = 403;
+                throw error;
+            }
             //check
             clearImage(post.imageUrl);
             return Post.findByIdAndRemove(postId);
         })
-        .then(result => {
-            console.log(result);
+        .then(() => {
+            return User.findById(req.userId);
+        })
+        .then(user => {
+            user.posts.pull(postId);
+            return user.save();
+        })
+        .then(() => {
             res
                 .status(200)
                 .json({
                     message: 'Deleted post.'
-            });
+                });
         })
         .catch(err => {
             if (!err.statusCode) {
